requirements.txt


Pillow==8.0.1
customtkinter==5.2.2

